LFS-18: As an admin, I can specify which users can access a certain variable Move the permissions manager to the core permissions module
